test(card): one store spec, split by what each file pins
The two files that both tested the store are now one set, split by
describe: store.test.ts holds the store's own surface, and filters,
facets, subscriptions, events and the degraded stack each have a file
named for what they pin, beside the day clock's.

Every case keeps its name and its describe, and its body moves
unchanged. The two harnesses become one: every store is built with the
zero backoff the revamp spec used, and flush runs one round by default,
so the cases that were written against a two-round flush ask for two.

Closes #231
