+++ cards/haventory-card/src/store/store.test.ts
@@ -1,16 +1,38 @@
 import { describe, it, expect, vi } from 'vitest';
-import { Store } from './store';
+import { BULK_CHUNK_SIZE, Store, makeBulkOp } from './store';
 import { makeMockHass, makeItem } from '../test.utils';
-import type { StoreState } from './types';
+import type { Location, StoreState } from './types';
 
 /**
- * Let queued microtasks and any zero-delay timers run.
+ * The store's own surface: what `init` loads, what an optimistic write does to
+ * the list before and after the answer, and what the caches hold.
  *
- * What the waits below are actually waiting for is a promise chain — a list or a
- * facet call answered by the mock — so draining the queue is the whole wait, and
- * a real delay was only ever a hedge on the scheduler.
+ * A store case belongs in the sibling file named for what it pins — filters,
+ * facets, subscriptions, events, the degraded stack, the day clock — and here
+ * when what it pins is the store itself.
  */
-async function flush(rounds = 2): Promise<void> {
+
+function loc(id: string, name: string, parentId: string | null = null): Location {
+  const displayPath = parentId ? `${parentId} / ${name}` : name;
+  return {
+    id,
+    name,
+    parent_id: parentId,
+    area_id: null,
+    path: {
+      id_path: parentId ? [parentId, id] : [id],
+      name_path: parentId ? [parentId, name] : [name],
+      display_path: displayPath,
+      sort_key: displayPath.toLowerCase(),
+    },
+  };
+}
+
+/** Stores under test never wait on real backoff. */
+const fast = { retryBaseMs: 0 };
+
+/** Let queued microtasks and any zero-delay timers run. */
+async function flush(rounds = 1): Promise<void> {
   for (let i = 0; i < rounds; i++) await new Promise((resolve) => setTimeout(resolve, 0));
 }
 
@@ -18,7 +40,7 @@ describe('Store', () => {
   it('initializes with stats, areas, locations, and first page of items', async () => {
     const items = Array.from({ length: 30 }, (_, i) => makeItem({ id: `${i}`, name: `Item ${i}` }));
     const hass = makeMockHass({ items });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
 
     expect(store.state.value.statsCounts).toBeTruthy();
@@ -29,23 +51,9 @@ describe('Store', () => {
     expect(store.state.value.connected.stats).toBe(true);
   });
 
-  it('debounces search and lists items; prefetch at ~70%', async () => {
-    const items = Array.from({ length: 80 }, (_, i) => makeItem({ id: `${i}`, name: `Item ${i}` }));
-    const hass = makeMockHass({ items });
-    const store = new Store(hass);
-    await store.init();
-    expect(store.state.value.items.length).toBeGreaterThan(0);
-    // First page default 50
-    expect(store.state.value.items.length).toBe(50);
-    await store.prefetchIfNeeded(0.69);
-    expect(store.state.value.items.length).toBe(50);
-    await store.prefetchIfNeeded(0.71);
-    expect(store.state.value.items.length).toBe(80);
-  });
-
   it('creates item and updates state optimistically', async () => {
     const hass = makeMockHass({ items: [] });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
 
     await store.createItem({ name: 'New Item', quantity: 5 });
@@ -57,7 +65,7 @@ describe('Store', () => {
   it('updates item optimistically and reconciles on success', async () => {
     const base = makeItem({ id: '1', name: 'Original', quantity: 10 });
     const hass = makeMockHass({ items: [base] });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
 
     await store.updateItem('1', { name: 'Updated', quantity: 20 });
@@ -69,7 +77,7 @@ describe('Store', () => {
   it('optimistic update on updateItem; conflict surfaces banner with actions', async () => {
     const base = makeItem({ id: '1', name: 'A' });
     const hass = makeMockHass({ items: [base], conflictOnUpdate: true });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
     // Trigger conflict
     await store.updateItem('1', { name: 'A2' }).catch(() => undefined);
@@ -90,7 +98,7 @@ describe('Store', () => {
   it('rolls back optimistic changes when updateItem fails', async () => {
     const base = makeItem({ id: '1', name: 'Original', quantity: 1 });
     const hass = makeMockHass({ items: [base], conflictOnUpdate: true });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
 
     await store.updateItem('1', { name: 'Changed', quantity: 5 }).catch(() => undefined);
@@ -104,7 +112,7 @@ describe('Store', () => {
   it('deletes item optimistically and rolls back on error', async () => {
     const item = makeItem({ id: '1', name: 'ToDelete' });
     const hass = makeMockHass({ items: [item] });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
 
     expect(store.state.value.items.length).toBe(1);
@@ -115,7 +123,7 @@ describe('Store', () => {
   it('adjusts quantity with optimistic update', async () => {
     const item = makeItem({ id: '1', quantity: 10 });
     const hass = makeMockHass({ items: [item] });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
 
     // Optimistic update happens immediately
@@ -129,7 +137,7 @@ describe('Store', () => {
   it('sets quantity with optimistic update', async () => {
     const item = makeItem({ id: '1', quantity: 10 });
     const hass = makeMockHass({ items: [item] });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
 
     // Optimistic update happens immediately
@@ -142,7 +150,7 @@ describe('Store', () => {
   it('checks out and checks in items with optimistic updates', async () => {
     const item = makeItem({ id: '1', checked_out: false });
     const hass = makeMockHass({ items: [item] });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
 
     // Check out optimistically
@@ -162,7 +170,7 @@ describe('Store', () => {
   it('sets low stock threshold with optimistic update', async () => {
     const item = makeItem({ id: '1', low_stock_threshold: null });
     const hass = makeMockHass({ items: [item] });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
 
     // Set threshold optimistically
@@ -182,7 +190,7 @@ describe('Store', () => {
     // Optimistic update: location_id changes immediately before server response
     const item = makeItem({ id: '1', location_id: 'loc1' });
     const hass = makeMockHass({ items: [item] });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
 
     // Start move - optimistic update should happen immediately
@@ -196,58 +204,23 @@ describe('Store', () => {
     expect(moved?.location_id).toBe('loc2');
   });
 
-  it('handles filter changes and restarts paging', async () => {
-    const items = Array.from({ length: 30 }, (_, i) => makeItem({ id: `${i}`, name: `Item ${i}` }));
-    const hass = makeMockHass({ items });
-    const store = new Store(hass);
-    await store.init();
-
-    const initialCount = store.state.value.items.length;
-    expect(initialCount).toBe(30);
-
-    store.setFilters({ q: 'search term' });
-    expect(store.state.value.filters.q).toBe('search term');
-    expect(store.state.value.cursor).toBe(null);
-  });
-
-  // Blanking the list here is what tore the card's scroller down mid-edit and
-  // took the open editor with it. The refetch replaces the rows when it lands.
-  it('keeps the loaded rows and the total while a filter refetch is in flight', async () => {
-    const items = Array.from({ length: 3 }, (_, i) => makeItem({ id: `${i}`, name: `Item ${i}` }));
-    const hass = makeMockHass({ items });
-    const store = new Store(hass);
-    await store.init();
-    const before = store.state.value.items;
-    const total = store.state.value.total;
-    expect(total).toBe(3);
-
-    store.setFilters({ q: 'Item 1' });
-    expect(store.state.value.items).toBe(before);
-    expect(store.state.value.total).toBe(total);
-    expect(store.state.value.loading).toBe(true);
-
-    await flush();
-    expect(store.state.value.loading).toBe(false);
-    expect(store.state.value.items.map((i) => i.name)).toEqual(['Item 1']);
-  });
-
   // Two disappearances look identical from the item list: a row that fell off
   // the filtered page, and one that is gone. Only the second closes an editor.
   describe('wasRemoved', () => {
     it('reports nothing removed for a row that is merely filtered out', async () => {
       const hass = makeMockHass({ items: [makeItem({ id: 'a', name: 'Alpha' }), makeItem({ id: 'b', name: 'Beta' })] });
-      const store = new Store(hass);
+      const store = new Store(hass, fast);
       await store.init();
 
       store.setFilters({ q: 'Alpha' });
-      await flush();
+      await flush(2);
       expect(store.state.value.items.map((i) => i.id)).toEqual(['a']);
       expect(store.wasRemoved('b')).toBe(false);
     });
 
     it('reports a deleted row as removed', async () => {
       const hass = makeMockHass({ items: [makeItem({ id: 'a' })] });
-      const store = new Store(hass);
+      const store = new Store(hass, fast);
       await store.init();
 
       await store.deleteItem('a', 1);
@@ -256,7 +229,7 @@ describe('Store', () => {
 
     it('forgets a delete that was rolled back', async () => {
       const hass = makeMockHass({ items: [makeItem({ id: 'a' })] });
-      const store = new Store(hass, { retryBaseMs: 0 });
+      const store = new Store(hass, fast);
       await store.init();
       const passthrough = hass.callWS.bind(hass);
       hass.callWS = async <T,>(msg: Record<string, unknown>): Promise<T> => {
@@ -273,7 +246,7 @@ describe('Store', () => {
   it('dismisses errors from error queue', async () => {
     const item = makeItem({ id: '1' });
     const hass = makeMockHass({ items: [item], conflictOnUpdate: true });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
 
     // Trigger an error
@@ -288,7 +261,7 @@ describe('Store', () => {
   it('refreshes single item from backend', async () => {
     const item = makeItem({ id: '1', name: 'Original' });
     const hass = makeMockHass({ items: [item] });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
 
     // refreshItem should fetch latest version
@@ -300,7 +273,7 @@ describe('Store', () => {
 
   it('createLocation calls WS and refreshes caches', async () => {
     const hass = makeMockHass({ items: [] });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
 
     const created = await store.createLocation('Test Location', null);
@@ -313,7 +286,7 @@ describe('Store', () => {
 
   it('updates stats cache on refreshStats', async () => {
     const hass = makeMockHass({ items: [] });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
 
     expect(store.state.value.statsCounts).toBeTruthy();
@@ -323,85 +296,9 @@ describe('Store', () => {
     expect(typeof store.state.value.statsCounts?.items_total).toBe('number');
   });
 
-  it('sends orphaned_only to the backend when the orphans filter is on', async () => {
-    const hass = makeMockHass({ items: [] });
-    const store = new Store(hass);
-    await store.init();
-
-    const listFilters: any[] = [];
-    const origCallWS = hass.callWS.bind(hass);
-    hass.callWS = async <T,>(msg: Record<string, unknown>): Promise<T> => {
-      if (msg.type === 'haventory/item/list') listFilters.push(msg.filter);
-      return origCallWS<T>(msg);
-    };
-
-    store.setFilters({ orphansOnly: true });
-    await flush();
-    expect(listFilters.length).toBeGreaterThan(0);
-    expect(listFilters[listFilters.length - 1]?.orphaned_only).toBe(true);
-
-    store.setFilters({ orphansOnly: false });
-    await flush();
-    expect(listFilters[listFilters.length - 1]?.orphaned_only).toBeUndefined();
-  });
-
-  it('orphans filter narrows the visible items end-to-end through the mock', async () => {
-    const placed = makeItem({
-      id: 'p1', name: 'Placed', location_id: 'loc1',
-      location_path: { id_path: ['loc1'], name_path: ['Garage'], display_path: 'Garage', sort_key: 'garage' },
-    });
-    const orphan = makeItem({ id: 'o1', name: 'Orphan' });
-    const hass = makeMockHass({ items: [placed, orphan] });
-    const store = new Store(hass);
-    await store.init();
-    expect(store.state.value.items.length).toBe(2);
-
-    store.setFilters({ orphansOnly: true });
-    await flush();
-    expect(store.state.value.items.map((i) => i.id)).toEqual(['o1']);
-
-    store.setFilters({ orphansOnly: false });
-    await flush();
-    expect(store.state.value.items.length).toBe(2);
-  });
-
-  it('q filter narrows the visible items end-to-end through the mock', async () => {
-    const a = makeItem({ id: 'a', name: 'Electric Saw' });
-    const b = makeItem({ id: 'b', name: 'Glue', tags: ['adhesive'] });
-    const hass = makeMockHass({ items: [a, b] });
-    const store = new Store(hass);
-    await store.init();
-
-    store.setFilters({ q: 'saw' });
-    await flush();
-    expect(store.state.value.items.map((i) => i.name)).toEqual(['Electric Saw']);
-
-    // Tags are searchable too
-    store.setFilters({ q: 'adhesive' });
-    await flush();
-    expect(store.state.value.items.map((i) => i.name)).toEqual(['Glue']);
-  });
-
-  it('due_date sort orders items through the mock, undated last in both orders', async () => {
-    const early = makeItem({ id: 'e', name: 'Early', checked_out: true, due_date: '2024-01-01' });
-    const late = makeItem({ id: 'l', name: 'Late', checked_out: true, due_date: '2024-03-01' });
-    const undated = makeItem({ id: 'u', name: 'Undated' });
-    const hass = makeMockHass({ items: [undated, late, early] });
-    const store = new Store(hass);
-    await store.init();
-
-    store.setFilters({ sort: { field: 'due_date', order: 'asc' } });
-    await flush();
-    expect(store.state.value.items.map((i) => i.name)).toEqual(['Early', 'Late', 'Undated']);
-
-    store.setFilters({ sort: { field: 'due_date', order: 'desc' } });
-    await flush();
-    expect(store.state.value.items.map((i) => i.name)).toEqual(['Late', 'Early', 'Undated']);
-  });
-
   it('populates healthCache on init and refreshes on demand', async () => {
     const hass = makeMockHass({ items: [makeItem({ id: '1', name: 'A' })] });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
 
     expect(store.state.value.healthCache?.counts.items_total).toBe(1);
@@ -413,7 +310,7 @@ describe('Store', () => {
 
   it('deleteLocation removes an empty location and refreshes caches', async () => {
     const hass = makeMockHass({ items: [] });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
 
     const created = await store.createLocation('Doomed', null);
@@ -425,7 +322,7 @@ describe('Store', () => {
 
   it('deleteLocation rejects with validation_error when the location is not empty', async () => {
     const hass = makeMockHass({ items: [] });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
 
     const parent = await store.createLocation('Parent', null);
@@ -438,7 +335,7 @@ describe('Store', () => {
 
   it('deleteLocation rejects when items still reference the location', async () => {
     const hass = makeMockHass({ items: [] });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
 
     const loc = await store.createLocation('Occupied', null);
@@ -449,7 +346,7 @@ describe('Store', () => {
 
   it('moveLocationSubtree reparents and refreshes locations and items', async () => {
     const hass = makeMockHass({ items: [makeItem({ id: 'i1', name: 'Inside' })] });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
 
     const src = await store.createLocation('Source', null);
@@ -459,486 +356,6 @@ describe('Store', () => {
     expect(moved.parent_id).toBe(dst.id);
     const cached = store.state.value.locationsFlatCache?.find((l) => l.id === src.id);
     expect(cached?.parent_id).toBe(dst.id);
-  });
-
-  it('reloads the item list when a locations moved/renamed event arrives', async () => {
-    const hass = makeMockHass({ items: [makeItem({ id: 'i1', name: 'Widget' })] });
-    const store = new Store(hass);
-    await store.init();
-
-    // Count item/list calls from here on via the callWS wrapper
-    let listCalls = 0;
-    const origCallWS = hass.callWS.bind(hass);
-    hass.callWS = async <T,>(msg: Record<string, unknown>): Promise<T> => {
-      if (msg.type === 'haventory/item/list') listCalls += 1;
-      return origCallWS<T>(msg);
-    };
-
-    const loc = {
-      id: 'locX', parent_id: null, name: 'Moved', area_id: null,
-      path: { id_path: ['locX'], name_path: ['Moved'], display_path: 'Moved', sort_key: 'moved' },
-    };
-    hass.__emit('locations', 'moved', { location: loc });
-    await new Promise((r) => setTimeout(r, 0));
-    expect(listCalls).toBe(1);
-
-    hass.__emit('locations', 'renamed', { location: loc });
-    await new Promise((r) => setTimeout(r, 0));
-    expect(listCalls).toBe(2);
-
-    // Non-path-changing events do not reload the list
-    hass.__emit('locations', 'created', { location: loc });
-    await new Promise((r) => setTimeout(r, 0));
-    expect(listCalls).toBe(2);
-  });
-
-  it('updates statsCounts live from a stats counts event', async () => {
-    const hass = makeMockHass({ items: [] });
-    const store = new Store(hass);
-    await store.init();
-
-    const counts = { items_total: 42, low_stock_count: 3, checked_out_count: 1, locations_total: 7 };
-    hass.__emit('stats', 'counts', { counts });
-    await new Promise((r) => setTimeout(r, 0));
-    expect(store.state.value.statsCounts).toEqual(expect.objectContaining({ items_total: 42, locations_total: 7 }));
-  });
-
-  it('caches distinct categories and tags with counts on init', async () => {
-    const items = [
-      makeItem({ id: '1', category: 'Tools', tags: ['red'] }),
-      makeItem({ id: '2', category: 'Tools', tags: ['red', 'blue'] }),
-      makeItem({ id: '3', category: 'Books', tags: ['blue'] }),
-    ];
-    const hass = makeMockHass({ items });
-    const store = new Store(hass);
-    await store.init();
-
-    const distinct = store.state.value.distinctValuesCache;
-    expect(distinct).toBeTruthy();
-    expect(distinct!.categories).toEqual([
-      { value: 'Books', count: 1 },
-      { value: 'Tools', count: 2 },
-    ]);
-    expect(distinct!.tags).toEqual([
-      { value: 'blue', count: 2 },
-      { value: 'red', count: 2 },
-    ]);
-  });
-
-  it('refreshes distinct values after an item is created', async () => {
-    const hass = makeMockHass({ items: [] });
-    const store = new Store(hass);
-    await store.init();
-    expect(store.state.value.distinctValuesCache?.categories).toEqual([]);
-
-    await store.createItem({ name: 'Drill', category: 'Tools' });
-    // The items subscription event triggers a distinct-values refresh.
-    hass.__emit('items', 'created', {
-      item: makeItem({ id: '99', name: 'Drill', category: 'Tools' }),
-    });
-    await new Promise((r) => setTimeout(r, 0));
-
-    expect(store.state.value.distinctValuesCache?.categories).toEqual([
-      { value: 'Tools', count: 1 },
-    ]);
-  });
-
-  // A category or tag exists only through the items carrying it — the backend
-  // has no registry to create an empty one in. A value the user names up front
-  // is therefore kept on the card until an item adopts it.
-  it('lists a newly named category alongside the ones in use, at zero', async () => {
-    const store = new Store(makeMockHass({ items: [makeItem({ id: '1', category: 'Tools' })] }));
-    await store.init();
-
-    expect(store.addDraftValue('category', 'Consumables')).toBe(true);
-    expect(store.state.value.distinctValuesCache?.categories).toEqual([
-      { value: 'Consumables', count: 0 },
-      { value: 'Tools', count: 1 },
-    ]);
-  });
-
-  it('normalizes a new tag the way the backend would', async () => {
-    const store = new Store(makeMockHass({ items: [] }));
-    await store.init();
-
-    store.addDraftValue('tag', '  Power Tools  ');
-    expect(store.state.value.distinctValuesCache?.tags).toEqual([{ value: 'power tools', count: 0 }]);
-  });
-
-  it('refuses a name that already exists, whatever its casing', async () => {
-    const store = new Store(makeMockHass({ items: [makeItem({ id: '1', category: 'Tools' })] }));
-    await store.init();
-
-    expect(store.addDraftValue('category', 'tools')).toBe(false);
-    expect(store.addDraftValue('category', '   ')).toBe(false);
-    expect(store.state.value.distinctValuesCache?.categories).toEqual([{ value: 'Tools', count: 1 }]);
-  });
-
-  it('stops carrying the draft once an item actually uses it', async () => {
-    const hass = makeMockHass({ items: [] });
-    const store = new Store(hass);
-    await store.init();
-    store.addDraftValue('category', 'Consumables');
-
-    await store.createItem({ name: 'Sponges', category: 'Consumables' });
-    hass.__emit('items', 'created', {
-      item: makeItem({ id: '99', name: 'Sponges', category: 'Consumables' }),
-    });
-    await new Promise((r) => setTimeout(r, 0));
-
-    // One entry, with the server's count — not a duplicate sitting at zero.
-    expect(store.state.value.distinctValuesCache?.categories).toEqual([
-      { value: 'Consumables', count: 1 },
-    ]);
-  });
-
-  it('discards a draft again', async () => {
-    const store = new Store(makeMockHass({ items: [] }));
-    await store.init();
-    store.addDraftValue('tag', 'seasonal');
-    store.removeDraftValue('tag', 'seasonal');
-
-    expect(store.state.value.distinctValuesCache?.tags).toEqual([]);
-  });
-
-  // The Location column's header is a sort control now, so the whole round trip
-  // — field on the wire, ordered page back — has to hold end to end.
-  it('lists items ordered by their location path, unlocated last', async () => {
-    const path = (display: string, ids: string[]) => ({
-      id_path: ids,
-      name_path: display.split(' / '),
-      display_path: display,
-      sort_key: display.toLowerCase(),
-    });
-    const items = [
-      makeItem({ id: '1', name: 'Deep', location_id: 'shelf', location_path: path('Garage / Shelf A', ['garage', 'shelf']) }),
-      makeItem({ id: '2', name: 'Loose' }),
-      makeItem({ id: '3', name: 'Elsewhere', location_id: 'cellar', location_path: path('Cellar', ['cellar']) }),
-      makeItem({ id: '4', name: 'Shallow', location_id: 'garage', location_path: path('Garage', ['garage']) }),
-    ];
-    const hass = makeMockHass({ items });
-    const store = new Store(hass);
-    await store.init();
-
-    store.setFilters({ sort: { field: 'location', order: 'asc' } });
-    await vi.waitUntil(() => store.state.value.items[0]?.name === 'Elsewhere');
-    expect(store.state.value.items.map((i) => i.name)).toEqual([
-      'Elsewhere',
-      'Shallow',
-      'Deep',
-      'Loose',
-    ]);
-
-    // Reversed, the located items flip and the unlocated one stays at the end.
-    store.setFilters({ sort: { field: 'location', order: 'desc' } });
-    await vi.waitUntil(() => store.state.value.items[0]?.name === 'Deep');
-    expect(store.state.value.items.map((i) => i.name)).toEqual([
-      'Deep',
-      'Shallow',
-      'Elsewhere',
-      'Loose',
-    ]);
-
-    const sent = hass.__messages.filter((m) => m.type === 'haventory/item/list');
-    expect(sent[sent.length - 1].sort).toEqual({ field: 'location', order: 'desc' });
-  });
-
-  it('asks for unpriced facets while nothing is filtering', async () => {
-    const hass = makeMockHass({ items: [makeItem({ id: '1', category: 'Tools', tags: ['red'] })] });
-    const store = new Store(hass);
-    await store.init();
-
-    const sent = hass.__messages.filter((m) => m.type === 'haventory/distinct_values');
-    expect(sent).toHaveLength(1);
-    expect(sent[0].filter).toBeUndefined();
-    // No key at all rather than a matching_count equal to count: the two say
-    // different things, and only one of them is a report.
-    expect(store.state.value.distinctValuesCache?.categories).toEqual([
-      { value: 'Tools', count: 1 },
-    ]);
-  });
-
-  it('prices the facets against the active filter, minus category and tags', async () => {
-    const items = [
-      makeItem({ id: '1', category: 'Tools', tags: ['red'], checked_out: true }),
-      makeItem({ id: '2', category: 'Tools', tags: ['red'] }),
-      makeItem({ id: '3', category: 'Books', tags: ['blue'] }),
-    ];
-    const hass = makeMockHass({ items });
-    const store = new Store(hass);
-    await store.init();
-
-    // A category and a tag are picked alongside the narrowing filter; neither
-    // may reach the wire, or every other row would be priced at zero.
-    store.setFilters({ checkedOutOnly: true, categories: ['Tools'], tags: ['red'] });
-    await vi.waitUntil(
-      () => hass.__messages.filter((m) => m.type === 'haventory/distinct_values').length > 1,
-    );
-
-    const sent = hass.__messages.filter((m) => m.type === 'haventory/distinct_values');
-    const filter = sent[sent.length - 1].filter as Record<string, unknown>;
-    expect(filter.checked_out).toBe(true);
-    expect(filter.category).toBeUndefined();
-    expect(filter.tags_any).toBeUndefined();
-
-    expect(store.state.value.distinctValuesCache?.categories).toEqual([
-      { value: 'Books', count: 1, matching_count: 0 },
-      { value: 'Tools', count: 2, matching_count: 1 },
-    ]);
-  });
-
-  // The commonest filter of all is a category or a tag on its own. Gating the
-  // pair on what survives `facetCountFilters` left exactly that case mixed:
-  // location rows reading "8 / 37" beside category rows reading "43".
-  it('prices every list when the only filter is one the facets drop', async () => {
-    const items = [
-      makeItem({ id: '1', category: 'Tools', tags: ['red'] }),
-      makeItem({ id: '2', category: 'Books', tags: ['blue'] }),
-    ];
-    const hass = makeMockHass({ items });
-    const store = new Store(hass);
-    await store.init();
-
-    store.setFilters({ categories: ['Tools'] });
-    await vi.waitUntil(
-      () => store.state.value.distinctValuesCache?.categories[0]?.matching_count !== undefined,
-    );
-
-    // Nothing else is narrowing, so every row prices at n / n — true, and the
-    // same shape the location rows are showing at that moment.
-    expect(store.state.value.distinctValuesCache?.categories).toEqual([
-      { value: 'Books', count: 1, matching_count: 1 },
-      { value: 'Tools', count: 1, matching_count: 1 },
-    ]);
-
-    const sent = hass.__messages.filter((m) => m.type === 'haventory/distinct_values');
-    expect(sent[sent.length - 1].filter).toBeDefined();
-  });
-
-  // The same asymmetry the other way round: a lone location filter must not
-  // leave the tree bare while the facet lists beside it carry a pair.
-  it('prices the tree when the only filter is the one it drops', async () => {
-    const hass = makeMockHass({
-      items: [makeItem({ id: '1', location_id: 'garage' })],
-      locations: [
-        {
-          id: 'garage',
-          name: 'Garage',
-          parent_id: null,
-          area_id: null,
-          path: { id_path: ['garage'], name_path: ['Garage'], display_path: 'Garage', sort_key: 'garage' },
-        },
-      ],
-    });
-    const store = new Store(hass);
-    await store.init();
-
-    store.setFilters({ locationIds: ['garage'] });
-    await vi.waitUntil(() => store.state.value.locationTreeCache?.[0]?.matching_subtree_count !== undefined);
-
-    expect(store.state.value.locationTreeCache?.[0].matching_subtree_count).toBe(1);
-    expect(store.state.value.locationMatchTotal).toBe(1);
-  });
-
-  it('coalesces the facet refetch across a burst of filter patches', async () => {
-    const hass = makeMockHass({ items: [makeItem({ id: '1', category: 'Tools' })] });
-    const store = new Store(hass);
-    await store.init();
-    const before = hass.__messages.filter((m) => m.type === 'haventory/distinct_values').length;
-
-    // What a filter panel does: several keys in a row, one answer wanted.
-    // The coalescing window is a 250 ms debounce, so the test drives that clock
-    // rather than out-waiting it: installed here, after init, so the setup above
-    // keeps running on real timers.
-    vi.useFakeTimers({ toFake: ['setTimeout', 'clearTimeout'] });
-    try {
-      store.setFilters({ checkedOutOnly: true });
-      store.setFilters({ lowStockOnly: true });
-      store.setFilters({ q: 'drill' });
-
-      // Nothing has been asked yet on the last millisecond of the window...
-      await vi.advanceTimersByTimeAsync(249);
-      const facetCalls = () =>
-        hass.__messages.filter((m) => m.type === 'haventory/distinct_values').length;
-      expect(facetCalls()).toBe(before);
-
-      // ...and the three patches then produce exactly one answer, with nothing
-      // trailing behind it.
-      await vi.advanceTimersByTimeAsync(1);
-      expect(facetCalls()).toBe(before + 1);
-      await vi.advanceTimersByTimeAsync(300);
-      expect(facetCalls()).toBe(before + 1);
-    } finally {
-      vi.useRealTimers();
-    }
-  });
-
-  // Not every facet refetch is debounced — an item event lands
-  // beside a filter change — so two can be in flight at once, and the response
-  // that lands last is not the one that was issued last. The newest request is
-  // the only one allowed to assign.
-  it('drops a facet response from a superseded request', async () => {
-    const hass = makeMockHass({ items: [makeItem({ id: '1', category: 'Fresh' })] });
-    const store = new Store(hass);
-    await store.init();
-
-    const original = hass.callWS.bind(hass);
-    let release: (() => void) | null = null;
-    let intercepted = false;
-    hass.callWS = (async (msg: Record<string, unknown>) => {
-      if (String(msg.type) === 'haventory/distinct_values' && !intercepted) {
-        intercepted = true;
-        // The older request: answer late, with data the inventory no longer holds.
-        await new Promise<void>((resolve) => {
-          release = resolve;
-        });
-        return { categories: [{ value: 'Stale', count: 9 }], tags: [], custom_field_keys: [] };
-      }
-      return original(msg);
-    }) as typeof hass.callWS;
-
-    const first = store.refreshDistinctValues();
-    const second = store.refreshDistinctValues();
-    await second;
-    expect(store.state.value.distinctValuesCache?.categories).toEqual([
-      { value: 'Fresh', count: 1 },
-    ]);
-
-    await vi.waitUntil(() => release !== null);
-    release!();
-    await first;
-    // The late answer belongs to a superseded request and must not land.
-    expect(store.state.value.distinctValuesCache?.categories).toEqual([
-      { value: 'Fresh', count: 1 },
-    ]);
-  });
-
-  // The per-location counts ride the tree, so it has the same overlap and gets
-  // the same guard.
-  it('drops a tree response from a superseded refetch', async () => {
-    const garage = {
-      id: 'garage',
-      name: 'Garage',
-      parent_id: null,
-      area_id: null,
-      path: { id_path: ['garage'], name_path: ['Garage'], display_path: 'Garage', sort_key: 'garage' },
-    };
-    const hass = makeMockHass({ items: [], locations: [garage] });
-    const store = new Store(hass);
-    await store.init();
-
-    const original = hass.callWS.bind(hass);
-    let release: (() => void) | null = null;
-    let intercepted = false;
-    hass.callWS = (async (msg: Record<string, unknown>) => {
-      if (String(msg.type) === 'haventory/location/tree' && !intercepted) {
-        intercepted = true;
-        await new Promise<void>((resolve) => {
-          release = resolve;
-        });
-        return [
-          {
-            id: 'stale',
-            name: 'Stale',
-            parent_id: null,
-            area_id: null,
-            path: { id_path: ['stale'], name_path: ['Stale'], display_path: 'Stale', sort_key: 'stale' },
-            direct_item_count: 0,
-            subtree_item_count: 0,
-            children: [],
-          },
-        ];
-      }
-      return original(msg);
-    }) as typeof hass.callWS;
-
-    const first = store.refreshLocationTree();
-    const second = store.refreshLocationTree();
-    await second;
-    expect(store.state.value.locationTreeCache?.map((n) => n.name)).toEqual(['Garage']);
-
-    await vi.waitUntil(() => release !== null);
-    release!();
-    await first;
-    expect(store.state.value.locationTreeCache?.map((n) => n.name)).toEqual(['Garage']);
-  });
-
-  it('gives a draft the priced shape once the server priced the rest', async () => {
-    const hass = makeMockHass({ items: [makeItem({ id: '1', category: 'Tools', checked_out: true })] });
-    const store = new Store(hass);
-    await store.init();
-
-    store.setFilters({ checkedOutOnly: true });
-    await vi.waitUntil(
-      () => hass.__messages.filter((m) => m.type === 'haventory/distinct_values').length > 1,
-    );
-    store.addDraftValue('category', 'Consumables');
-
-    // A draft has no items, so it matches nothing — said in the same shape the
-    // priced rows use, rather than reading as an unpriced row among them.
-    expect(store.state.value.distinctValuesCache?.categories).toEqual([
-      { value: 'Consumables', count: 0, matching_count: 0 },
-      { value: 'Tools', count: 1, matching_count: 1 },
-    ]);
-  });
-
-  it('scopes the items subscription to the active area and re-opens it when the area changes', async () => {
-    const hass = makeMockHass({ items: [makeItem({ id: '1' })] });
-    const store = new Store(hass);
-    await store.init();
-
-    const itemsSubscribe = () => {
-      const sent = hass.__subscribeMessages.filter((m) => m.topic === 'items');
-      return sent[sent.length - 1];
-    };
-
-    // Unfiltered, the subscription asks for every area.
-    expect(itemsSubscribe().area_id).toBe(null);
-    const roundOne = hass.__subscribeCalls.length;
-
-    store.setFilters({ areaId: 'kitchen' });
-    await new Promise((r) => setTimeout(r, 0));
-
-    // A fresh round of all four topics, and the items one now names the area —
-    // without it the card would keep receiving every other area's events.
-    expect(hass.__subscribeCalls.length).toBe(roundOne * 2);
-    expect(itemsSubscribe().area_id).toBe('kitchen');
-
-    // A filter the backend applies to the page rather than to the subscription
-    // leaves the sockets alone.
-    store.setFilters({ categories: ['Tools'] });
-    await new Promise((r) => setTimeout(r, 0));
-    expect(hass.__subscribeCalls.length).toBe(roundOne * 2);
-  });
-
-  // The subscription is scoped by location as well, so a multi-select that only
-  // reached the page query would leave the socket delivering the other
-  // locations' events — and the card would act on them.
-  it('scopes the items subscription to every selected location', async () => {
-    const hass = makeMockHass({ items: [makeItem({ id: '1' })] });
-    const store = new Store(hass);
-    await store.init();
-    const itemsSubscribe = () => {
-      const sent = hass.__subscribeMessages.filter((m) => m.topic === 'items');
-      return sent[sent.length - 1];
-    };
-    const roundOne = hass.__subscribeCalls.length;
-
-    store.setFilters({ locationIds: ['garage', 'kitchen'] });
-    await new Promise((r) => setTimeout(r, 0));
-
-    expect(hass.__subscribeCalls.length).toBe(roundOne * 2);
-    expect(itemsSubscribe().location_ids).toEqual(['garage', 'kitchen']);
-
-    // Re-picking the same set changes no scope, so the sockets stay up.
-    store.setFilters({ locationIds: ['garage', 'kitchen'] });
-    await new Promise((r) => setTimeout(r, 0));
-    expect(hass.__subscribeCalls.length).toBe(roundOne * 2);
-
-    // Clearing the selection re-opens on every location.
-    store.setFilters({ locationIds: [] });
-    await new Promise((r) => setTimeout(r, 0));
-    expect(itemsSubscribe().location_ids).toEqual([]);
   });
 
   // Bumping is the one mutation with no optimistic update: where the next
@@ -952,7 +369,7 @@ describe('Store', () => {
       reminder_interval: { unit: 'days', count: 7 },
     });
     const hass = makeMockHass({ items: [item] });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
 
     await store.bumpReminder('1', item.version);
@@ -966,7 +383,7 @@ describe('Store', () => {
   it('surfaces a refused bump as an error and leaves the item alone', async () => {
     const item = makeItem({ id: '1', reminder_date: '2026-08-31', reminder_interval: null });
     const hass = makeMockHass({ items: [item] });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
 
     await store.bumpReminder('1', item.version);
@@ -984,7 +401,7 @@ describe('Store', () => {
    */
   it('keeps no register of writes in flight', async () => {
     const hass = makeMockHass({ items: [makeItem({ id: '1', name: 'A' })] });
-    const store = new Store(hass);
+    const store = new Store(hass, fast);
     await store.init();
 
     await store.updateItem('1', { name: 'B' });
@@ -992,5 +409,375 @@ describe('Store', () => {
     const noPendingOps: Extract<keyof StoreState, 'pendingOps'> extends never ? true : never = true;
     expect(noPendingOps).toBe(true);
     expect(Object.keys(store.state.value)).not.toContain('pendingOps');
+  });
+});
+
+describe('Store: selection', () => {
+  const items = [makeItem({ id: '1' }), makeItem({ id: '2' }), makeItem({ id: '3' })];
+
+  it('toggles, replaces and clears', async () => {
+    const store = new Store(makeMockHass({ items }), fast);
+    await store.init();
+
+    store.toggleSelected('1');
+    store.toggleSelected('2');
+    expect([...store.state.value.selection].sort()).toEqual(['1', '2']);
+
+    store.toggleSelected('1');
+    expect([...store.state.value.selection]).toEqual(['2']);
+
+    store.setSelected(['3']);
+    expect([...store.state.value.selection]).toEqual(['3']);
+
+    store.clearSelection();
+    expect(store.state.value.selection.size).toBe(0);
+  });
+
+  it('publishes a new Set so subscribers see the change', async () => {
+    const store = new Store(makeMockHass({ items }), fast);
+    await store.init();
+    const before = store.state.value.selection;
+    store.toggleSelected('1');
+    expect(store.state.value.selection).not.toBe(before);
+  });
+
+  it('select-all covers loaded rows only, and load-all extends it', async () => {
+    const many = Array.from({ length: 60 }, (_, i) => makeItem({ id: `i${i}` }));
+    const store = new Store(makeMockHass({ items: many }), fast);
+    await store.init();
+
+    store.selectAllLoaded();
+    expect(store.state.value.selection.size).toBe(50);
+    expect(store.state.value.total).toBe(60);
+
+    await store.loadAllThenSelectAll();
+    expect(store.state.value.selection.size).toBe(60);
+  });
+
+  it('drops the selection when the filter changes', async () => {
+    const store = new Store(makeMockHass({ items }), fast);
+    await store.init();
+    store.selectAllLoaded();
+    expect(store.state.value.selection.size).toBe(3);
+
+    store.setFilters({ q: 'nothing' });
+    expect(store.state.value.selection.size).toBe(0);
+  });
+});
+
+describe('Store: bulk operations', () => {
+  const three = [
+    makeItem({ id: '1', name: 'A', quantity: 5 }),
+    makeItem({ id: '2', name: 'B', quantity: 5 }),
+    makeItem({ id: '3', name: 'C', quantity: 5 }),
+  ];
+
+  it('applies every successful op and reports the items back', async () => {
+    const store = new Store(makeMockHass({ items: three }), fast);
+    await store.init();
+
+    const outcome = await store.bulkExecute(
+      three.map((i) => makeBulkOp('item_adjust_quantity', { item_id: i.id, delta: 2 })),
+    );
+
+    expect(outcome.failed).toEqual([]);
+    expect(outcome.cancelled).toBe(false);
+    expect(outcome.succeeded.map((i) => i.quantity)).toEqual([7, 7, 7]);
+    // The list reflects the batch without waiting for events.
+    expect(store.state.value.items.map((i) => i.quantity)).toEqual([7, 7, 7]);
+  });
+
+  it('separates the failures and names the row each belongs to', async () => {
+    const store = new Store(makeMockHass({ items: three }), fast);
+    await store.init();
+
+    const outcome = await store.bulkExecute([
+      makeBulkOp('item_adjust_quantity', { item_id: '1', delta: 1 }),
+      makeBulkOp('item_adjust_quantity', { item_id: 'ghost', delta: 1 }),
+      makeBulkOp('item_adjust_quantity', { item_id: '3', delta: 1 }),
+    ]);
+
+    expect(outcome.succeeded).toHaveLength(2);
+    expect(outcome.failed).toHaveLength(1);
+    expect(outcome.failed[0].itemId).toBe('ghost');
+    expect(outcome.failed[0].error.code).toBe('not_found');
+  });
+
+  it('gives every operation a unique op_id so results cannot collapse', () => {
+    const a = makeBulkOp('item_check_in', { item_id: '1' });
+    const b = makeBulkOp('item_check_in', { item_id: '1' });
+    expect(a.op_id).not.toBe(b.op_id);
+  });
+
+  it('chunks the work and reports determinate progress', async () => {
+    const many = Array.from({ length: 60 }, (_, i) => makeItem({ id: `i${i}`, quantity: 1 }));
+    const hass = makeMockHass({ items: many });
+    const store = new Store(hass, fast);
+    await store.init();
+    await store.loadAllPages();
+
+    const progress: [number, number][] = [];
+    await store.bulkExecute(
+      many.map((i) => makeBulkOp('item_adjust_quantity', { item_id: i.id, delta: 1 })),
+      { chunkSize: 25, onProgress: (done, total) => progress.push([done, total]) },
+    );
+
+    expect(progress).toEqual([
+      [25, 60],
+      [50, 60],
+      [60, 60],
+    ]);
+    expect(hass.__calls.filter((c) => c === 'haventory/items/bulk')).toHaveLength(3);
+  });
+
+  it('defaults to the documented chunk size', async () => {
+    const many = Array.from({ length: BULK_CHUNK_SIZE + 1 }, (_, i) => makeItem({ id: `i${i}` }));
+    const hass = makeMockHass({ items: many });
+    const store = new Store(hass, fast);
+    await store.init();
+
+    await store.bulkExecute(many.map((i) => makeBulkOp('item_check_in', { item_id: i.id })));
+    expect(hass.__calls.filter((c) => c === 'haventory/items/bulk')).toHaveLength(2);
+  });
+
+  it('stops between chunks when cancelled, keeping what already ran', async () => {
+    const many = Array.from({ length: 60 }, (_, i) => makeItem({ id: `i${i}`, quantity: 1 }));
+    const store = new Store(makeMockHass({ items: many }), fast);
+    await store.init();
+    await store.loadAllPages();
+
+    let chunks = 0;
+    const outcome = await store.bulkExecute(
+      many.map((i) => makeBulkOp('item_adjust_quantity', { item_id: i.id, delta: 1 })),
+      {
+        chunkSize: 25,
+        onProgress: () => {
+          chunks += 1;
+        },
+        isCancelled: () => chunks >= 1,
+      },
+    );
+
+    expect(outcome.cancelled).toBe(true);
+    // The first chunk completed and is not rolled back — this is not a transaction.
+    expect(outcome.succeeded).toHaveLength(25);
+  });
+
+  it('attributes a whole-call failure to every op in the chunk', async () => {
+    const hass = makeMockHass({ items: three });
+    const store = new Store(hass, fast);
+    await store.init();
+    hass.__failNext(1, { code: 'storage_error', message: 'disk full' });
+
+    const outcome = await store.bulkExecute(
+      three.map((i) => makeBulkOp('item_check_in', { item_id: i.id })),
+    );
+
+    expect(outcome.succeeded).toEqual([]);
+    expect(outcome.failed).toHaveLength(3);
+    expect(outcome.failed.every((f) => f.error.code === 'storage_error')).toBe(true);
+  });
+
+  it('removes successfully deleted rows from the list', async () => {
+    const store = new Store(makeMockHass({ items: three }), fast);
+    await store.init();
+
+    await store.bulkExecute([
+      makeBulkOp('item_delete', { item_id: '1' }),
+      makeBulkOp('item_delete', { item_id: '2' }),
+    ]);
+
+    expect(store.state.value.items.map((i) => i.id)).toEqual(['3']);
+  });
+});
+
+describe('Store: location tree and diagnostics data', () => {
+  it('exposes tree nodes with per-location counts', async () => {
+    const locations = [loc('garage', 'Garage'), loc('shelf-a', 'Shelf A', 'garage')];
+    const items = [
+      makeItem({ id: '1', location_id: 'garage' }),
+      makeItem({ id: '2', location_id: 'shelf-a' }),
+      makeItem({ id: '3', location_id: 'shelf-a' }),
+    ];
+    const store = new Store(makeMockHass({ items, locations }), fast);
+    await store.init();
+
+    const tree = store.state.value.locationTreeCache!;
+    expect(tree).toHaveLength(1);
+    expect(tree[0].id).toBe('garage');
+    expect(tree[0].direct_item_count).toBe(1);
+    expect(tree[0].subtree_item_count).toBe(3);
+    expect(tree[0].children[0].id).toBe('shelf-a');
+    expect(tree[0].children[0].subtree_item_count).toBe(2);
+  });
+
+  // Unfiltered, the tree is asked plainly — no filter key on the wire at all,
+  // so nodes come back without the matching pair.
+  it('leaves the tree unfiltered until something is actually narrowing', async () => {
+    const locations = [loc('garage', 'Garage')];
+    const store = new Store(makeMockHass({ items: [makeItem({ id: '1', location_id: 'garage' })], locations }), fast);
+    await store.init();
+
+    expect(store.state.value.locationTreeCache![0].matching_subtree_count).toBeUndefined();
+    expect(store.state.value.locationMatchTotal).toBe(null);
+  });
+
+  it('counts each location against the filter, ignoring the location filter itself', async () => {
+    const locations = [loc('garage', 'Garage'), loc('shelf-a', 'Shelf A', 'garage'), loc('kitchen', 'Kitchen')];
+    const items = [
+      makeItem({ id: '1', location_id: 'garage', category: 'Tools' }),
+      makeItem({ id: '2', location_id: 'shelf-a', category: 'Tools' }),
+      makeItem({ id: '3', location_id: 'shelf-a', category: 'Cables' }),
+      makeItem({ id: '4', location_id: 'kitchen', category: 'Tools' }),
+      makeItem({ id: '5', category: 'Tools' }),
+    ];
+    const hass = makeMockHass({ items, locations });
+    const store = new Store(hass, fast);
+    await store.init();
+    const treeCalls = () => hass.__calls.filter((c) => c === 'haventory/location/tree').length;
+    const before = treeCalls();
+
+    // The refresh rides a 250 ms debounce, so the test drives that clock rather
+    // than out-waiting it: installed here, after init, so the setup above keeps
+    // running on real timers.
+    vi.useFakeTimers({ toFake: ['setTimeout', 'clearTimeout'] });
+    try {
+      // Narrow by category *and* location: the sidebar still has to show where
+      // the other matches are, or picking a different branch becomes guesswork.
+      store.setFilters({ categories: ['Tools'], locationIds: ['kitchen'] });
+
+      // Nothing on the last millisecond before the window closes...
+      await vi.advanceTimersByTimeAsync(249);
+      expect(treeCalls()).toBe(before);
+
+      // ...and exactly one walk when it does.
+      await vi.advanceTimersByTimeAsync(1);
+      expect(treeCalls()).toBe(before + 1);
+    } finally {
+      vi.useRealTimers();
+    }
+
+    const tree = store.state.value.locationTreeCache!;
+    const garage = tree.find((n) => n.id === 'garage')!;
+    const kitchen = tree.find((n) => n.id === 'kitchen')!;
+    expect(garage.subtree_item_count).toBe(3);
+    expect(garage.matching_subtree_count).toBe(2);
+    expect(kitchen.matching_subtree_count).toBe(1);
+    // Four Tools items in total; the fourth has no location, and the sidebar
+    // derives that row from this number.
+    expect(store.state.value.locationMatchTotal).toBe(4);
+  });
+
+  it('does not re-walk the tree for a re-order, which changes no count', async () => {
+    const hass = makeMockHass({ items: [makeItem({ id: '1' })], locations: [loc('garage', 'Garage')] });
+    const store = new Store(hass, fast);
+    await store.init();
+
+    const before = hass.__calls.filter((c) => c === 'haventory/location/tree').length;
+    vi.useFakeTimers({ toFake: ['setTimeout', 'clearTimeout'] });
+    try {
+      store.setFilters({ sort: { field: 'name', order: 'asc' } });
+      // Well past the 250 ms debounce: the claim is that nothing was scheduled
+      // at all, not that it had not fired yet.
+      await vi.advanceTimersByTimeAsync(400);
+    } finally {
+      vi.useRealTimers();
+    }
+    expect(hass.__calls.filter((c) => c === 'haventory/location/tree').length).toBe(before);
+  });
+
+  it('caches the integration version for diagnostics', async () => {
+    const hass = makeMockHass({ items: [] });
+    const store = new Store(hass, fast);
+    await store.init();
+
+    expect(store.state.value.versionInfo?.integration_version).toBe('0.0.1');
+  });
+
+  it('reads the card heading configured in the integration', async () => {
+    const hass = makeMockHass({ items: [], cardTitle: 'Pantry' });
+    const store = new Store(hass, fast);
+    await store.init();
+
+    expect(store.state.value.cardTitle).toBe('Pantry');
+  });
+
+  it('reads the quick-filter pills configured in the integration', async () => {
+    const hass = makeMockHass({ items: [], quickFilters: ['low_stock', 'overdue'] });
+    const store = new Store(hass, fast);
+    await store.init();
+
+    expect(store.state.value.quickFilters).toEqual(['low_stock', 'overdue']);
+  });
+
+  // `null` and `[]` are different answers, and the store is where the two
+  // could most easily collapse into one falsy value.
+  it('keeps an explicit "no pills" apart from "no choice made"', async () => {
+    const none = new Store(makeMockHass({ items: [], quickFilters: [] }), fast);
+    await none.init();
+    expect(none.state.value.quickFilters).toEqual([]);
+
+    const unset = new Store(makeMockHass({ items: [], quickFilters: null }), fast);
+    await unset.init();
+    expect(unset.state.value.quickFilters).toBeNull();
+  });
+
+  it('drops a pill name this bundle does not know', async () => {
+    const hass = makeMockHass({ items: [], quickFilters: ['low_stock', 'sideways'] });
+    const store = new Store(hass, fast);
+    await store.init();
+
+    expect(store.state.value.quickFilters).toEqual(['low_stock']);
+  });
+
+  it('leaves the pills unset against a backend that does not report them', async () => {
+    const store = new Store(makeMockHass({ items: [] }), fast);
+    await store.init();
+
+    expect(store.state.value.quickFilters).toBeNull();
+  });
+
+  // The card bundle is served by the integration, so this only happens with a
+  // stale cached bundle — and a missing heading must not cost the user their
+  // items, which init loads after this call.
+  it('still loads when the backend does not know haventory/config', async () => {
+    const hass = makeMockHass({ items: [makeItem({ id: '1' })] });
+    const realCallWS = hass.callWS.bind(hass);
+    hass.callWS = (async (msg: Record<string, unknown>) => {
+      if (msg.type === 'haventory/config') throw { code: 'unknown_command', message: 'nope' };
+      return realCallWS(msg);
+    }) as typeof hass.callWS;
+
+    const store = new Store(hass, fast);
+    await store.init();
+
+    expect(store.state.value.cardTitle).toBeNull();
+    expect(store.state.value.items).toHaveLength(1);
+  });
+});
+
+describe('Store: export scopes', () => {
+  it('exports everything by default and the active filter on request', async () => {
+    const items = [
+      makeItem({ id: '1', name: 'Wood Glue', category: 'Adhesives' }),
+      makeItem({ id: '2', name: 'M4 Screws', category: 'Hardware' }),
+    ];
+    const hass = makeMockHass({ items });
+    const calls: unknown[] = [];
+    const originalCallWS = hass.callWS.bind(hass);
+    hass.callWS = async <T,>(msg: Record<string, unknown>): Promise<T> => {
+      if (msg.type === 'haventory/export') calls.push(msg.filter ?? null);
+      return originalCallWS<T>(msg);
+    };
+
+    const store = new Store(hass, fast);
+    await store.init();
+    store.setFilters({ categories: ['Hardware'] });
+
+    await store.exportDocument();
+    await store.exportDocument('view');
+
+    expect(calls[0]).toBe(null);
+    expect((calls[1] as { categories?: string[] }).categories).toEqual(['Hardware']);
   });
 });
